Lab final using STM32 ARM Cortex M4
Purpose of this lab final was to demonstrate learning from the entire semester in a course focused on low-level programming & computer architecture. In this lab we used bit-wise operators, PWM, timers, GPIO, and interrupts & ISR. We had to configure all of the registers to their proper values so that we could get the microcontroller to act as intended.